requirements.txt
fastmcp
llama-index-core
llama-index-llms-openai
duckduckgo_search
httpx
pydantic
airportsdata
python-dateutil
rich
openai
playwright
pillow